+++ app/src/main/java/og/ogstartracker/network/ArduinoApi.kt
@@ -8,7 +8,8 @@ interface ArduinoApi {
 
 	@GET("on")
 	suspend fun startSiderealTracking(
-		@Query("direction") direction: Int
+		@Query("direction") direction: Int,
+		@Query("trackingSpeed") tspeed: Int,
 	): Response<String>
 
 	@GET("off")