+++ app/src/main/java/og/ogstartracker/ui/screens/DashboardViewModel.kt
@@ -19,7 +19,7 @@ import og.ogstartracker.domain.events.ExpositionTesterEvent
 import og.ogstartracker.domain.events.PhotoControlEvent
 import og.ogstartracker.domain.events.SlewControlEvent
 import og.ogstartracker.domain.models.CheckListItem
-import og.ogstartracker.domain.models.Hemisphere
+import og.ogstartracker.domain.models.TrackingMode
 import og.ogstartracker.domain.usecases.arduino.StartCaptureUseCase
 import og.ogstartracker.domain.usecases.providers.DashboardUseCaseProvider
 import og.ogstartracker.domain.usecases.settings.SetNewSettingsUseCase
@@ -69,7 +69,8 @@ class DashboardViewModel internal constructor(
 		useCases.getSettings(SettingItem.EXPOSURE_TIME),
 		useCases.getSettings(SettingItem.DITHER_ACTIVE),
 		useCases.getSettings(SettingItem.SLEW_SPEED),
-	) { pixelSize, focalLength, exposureCount, exposureTime, ditherActive, slewSpeed ->
+		useCases.getSettings(SettingItem.STOP_TRACKING),
+	) { pixelSize, focalLength, exposureCount, exposureTime, ditherActive, slewSpeed, stopTracking ->
 		_uiState.update {
 			it.copy(
 				exposeTime = TextFieldState(text = exposureTime?.toString() ?: "", NotEmptyValidator()),
@@ -77,7 +78,8 @@ class DashboardViewModel internal constructor(
 				ditherFocalLength = TextFieldState(text = focalLength?.toString() ?: "", NotEmptyValidator()),
 				ditherPixelSize = TextFieldState(text = pixelSize?.toString() ?: "", NotEmptyValidator()),
 				slewValue = slewSpeed ?: 0,
-				ditheringEnabled = ditherActive == 1
+				ditheringEnabled = ditherActive == 1,
+				stopTrackingEnabled = stopTracking == 1,
 			)
 		}
 		return@combine
@@ -99,12 +101,10 @@ class DashboardViewModel internal constructor(
 	private val _uiState = MutableStateFlow(DashboardUiState())
 	val uiState = combine(
 		_uiState,
-		useCases.getCurrentHemisphereFlow(),
 		useCases.didUserSeeOnboarding(),
 		useCases.getLastArduinoMessage()
-	) { uiState, hemisphere, userSawOnboarding, lastMessage ->
+	) { uiState, userSawOnboarding, lastMessage ->
 		uiState.copy(
-			hemisphere = hemisphere,
 			shouldShowOnboardingDialog = !userSawOnboarding,
 			lastMessage = lastMessage
 		)
@@ -122,12 +122,12 @@ class DashboardViewModel internal constructor(
 	}
 
 	/**
-	 * This enables or disables sidereal tracking.
+	 * This enables or disables tracking.
 	 */
-	internal fun changeSidereal(active: Boolean) {
+	internal fun changeTracking(active: Boolean) {
 		if (active) {
 			sendCommand {
-				useCases.startSiderealTracking(uiState.value.hemisphere ?: return@sendCommand).onSuccess {
+				useCases.startSiderealTracking().onSuccess {
 					vibratorController.startVibrations(vibrationPatternThreeClick)
 					_uiState.update { it.copy(siderealActive = true) }
 				}
@@ -236,6 +236,19 @@ class DashboardViewModel internal constructor(
 						startPhotoCaptureTimer()
 						vibratorController.startVibrations(vibrationPatternThreeClick)
 					}
+				}
+			}
+
+			is PhotoControlEvent.StopTrackingActivation -> {
+				_uiState.update {
+					it.copy(
+						stopTrackingEnabled = photoControlEvent.active,
+					)
+				}
+				if (photoControlEvent.active) {
+					vibratorController.startVibrations(vibrationPatternThreeClick)
+				} else {
+					vibratorController.startVibrations(vibrationPatternClick)
 				}
 			}
 		}
@@ -405,6 +418,10 @@ class DashboardViewModel internal constructor(
 
 			stopPhotoCaptureTimer()
 			photoControlEvent(PhotoControlEvent.EndCapture)
+
+			if (uiState.value.stopTrackingEnabled) {
+				changeTracking(false)
+			}
 		}
 	}
 
@@ -511,16 +528,17 @@ class DashboardViewModel internal constructor(
 }
 
 data class DashboardUiState internal constructor(
-	val hemisphere: Hemisphere? = null,
 	val wifiConnected: Boolean = false,
 	val haveLocationPermission: Boolean = false,
 	val openedCheckbox: Boolean = false,
 	val siderealActive: Boolean = false,
 	val ditheringEnabled: Boolean = false,
+	val stopTrackingEnabled: Boolean = false,
 	val capturingActive: Boolean = false,
 	val exposureTestingActive: Boolean = false,
 	val lastMessage: String? = null,
 	val slewValue: Int = 0,
+	val trackingMode: TrackingMode = TrackingMode.SIDEREAL,
 	// photo control
 	val captureStartTime: Long? = null,
 	val captureCount: Int? = null,