+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## [1.0.5]
+
+- add multiple tracking modes (sidereal, solar and lunar)
+- fixed bug when hemisphere setting was ignored (sorry about that)
+- add option to disable tracking when intervalometer finishes
+- tracking info on main screen now shows current settings for hemisphere and tracking mode
+
+## [1.0.4]
+
+- add new exposition tester feature
+
 ## [1.0.4]
 
 - add new exposition tester feature